requirements.txt
anthropic>=0.40.0
discord.py>=2.4.0
beautifulsoup4>=4.12.0
requests>=2.31.0
supabase>=2.3.0
python-dotenv>=1.0.0
lxml>=4.9.0
aiohttp>=3.9.0
